requirements.txt
gymnasium
pybullet
tqdm
h5py
attrdict
scipy
scikit-image
#imutils
opencv-python-headless
trimesh
open3d
tensorboard
sb3-contrib
stable_baselines3
shapely==2.0.4
toppra==0.5.3
numpy
seaborn
klampt
pyfqmr
hdf5plugin
lightning
torchvision
wandb
voxelmap
PyMCubes
pybind11[global]==2.13.6
netifaces
pyside2
rospkg
hydra-core
ftfy
regex
tqdm
termcolor
PyOpenGL
